requirements.txt
Django==4.2.14
djangorestframework==3.15.2
requests==2.32.3
django-environ==0.10.0
psycopg2-binary==2.9.9
